requirements.txt
setuptools>18.5
jinja2
pyyaml
paramiko
pysnmp
